+++ Servers/Filters/vtkPointsProjectedHull.cxx
@@ -145,7 +145,7 @@ VTK_GETSIZECCWHULL(Z, 2);
 #define VTK_RECTANGLEINTERSECTION(which, dim) \
 int vtkPointsProjectedHull::rectangleIntersection##which(vtkPoints *R) \
 {                                                                      \
-  float bounds[6];                                             \
+  double bounds[6];                                            \
   R->Modified();                                               \
   R->GetBounds(bounds);                                        \
   double hmin, hmax, vmin, vmax;                               \
